feat: add a method for retrieving errors within the flux AST
The `GetErrors` method will traverse the AST and find any nodes that
implement the `error` interface. It will then add them to the list of
errors that were added while constructing the AST and these nodes should
also have information about where the error happened.

This also includes a way to limit the number of errors so that we can
implement a functionality to only print the first error (might be useful
for chronograf) or limit the command line output to the first 10 errors.
If a value of zero or lower is given, it will collect all errors.